Added conditions for octal and hex numbers
The commit includes logic for both %d and %i format specifiers,
as well as logic for handling decimal, octal and hex numbers.

Each of these numbers were considered if they were to be positive or
negative.

The function returns the position of the counter. I think this might be
changed.